Move file opened notifications to its own class.
